+++ firmware/src/webcfg.cpp
@@ -227,7 +227,7 @@ namespace {
         else if (preview == "signin")  screen_setup::showSignInChoice();
         else if (preview == "waiting") screen_setup::showPreparing();
         else if (preview == "email")   screen_setup::showEmailPairing("http://192.168.20.170");
-        else if (preview == "settings") screen_settings::showMenu("Atelier", "benoit@atome3d.com", 3, 6);
+        else if (preview == "settings") screen_settings::showMenu("Atelier", "benoit@atome3d.com", 3, 6, true, "1.6.0");
         else if (preview == "setwifi")  screen_settings::showWifi("Atelier", "192.168.20.170",
                                                                   WiFi.macAddress().c_str(), true);
         else if (preview == "setacct")  screen_settings::showAccount("benoit@atome3d.com", 6, true);